# Improve codegen for IsNan