This resolves https://github.com/dotnet/coreclr/issues/14846 by updating `IsNaN(x)` to just do `x != x` and improving the handling of `x != x` in the JIT to only require a single branch.

```
Total bytes of diff: -16325 (-0.37% of base)
    diff is an improvement.

Top file improvements by size (bytes):
      -16325 : System.Private.CoreLib.dasm (-0.37% of base)

1 total files with size differences (1 improved, 0 regressed), 0 unchanged.

Top method improvements by size (bytes):
       -2921 (-63.47% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,double,int):int
       -2918 (-61.30% of base) : System.Private.CoreLib.dasm - SpanHelpers:IndexOfAny(byref,double,double,double,int):int
       -1930 (-61.45% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,int):int
       -1923 (-61.44% of base) : System.Private.CoreLib.dasm - SpanHelpers:IndexOfAny(byref,double,double,int):int
        -934 (-56.44% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOf(byref,double,int):int

Top method improvements by size (percentage):
         -36 (-72.00% of base) : System.Private.CoreLib.dasm - Double:IsNaN(double):bool
         -80 (-65.04% of base) : System.Private.CoreLib.dasm - Double:Equals(double):bool:this
         -72 (-63.72% of base) : System.Private.CoreLib.dasm - GenericEqualityComparer`1:Equals(double,double):bool:this
       -2921 (-63.47% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,double,int):int
       -1930 (-61.45% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,int):int

57 total methods with size differences (57 improved, 0 regressed), 20447 unchanged.
```