@tannergooding 2019-10-17
`double.IsNaN` itself changes from:
```asm
G_M49729_IG01:
       push     rax
       vzeroupper 

G_M49729_IG02:
       vmovsd   qword ptr [rsp], xmm0
       mov      rax, qword ptr [rsp]
       mov      rdx, 0xD1FFAB1E
       and      rax, rdx
       mov      rdx, 0xD1FFAB1E
       cmp      rax, rdx
       setg     al
       movzx    rax, al

G_M49729_IG03:
       add      rsp, 8
       ret
```

To:
```asm
G_M49728_IG01:
       vzeroupper 

G_M49728_IG02:
       vucomisd xmm0, xmm0
       setp     al
       movzx    rax, al

G_M49728_IG03:
       ret
```