@tannergooding 2019-10-17
Also, for reference.... Just the System.Private.Corelib change gives the following diff
```
Total bytes of diff: -15134 (-0.34% of base)
    diff is an improvement.

Top file improvements by size (bytes):
      -15134 : System.Private.CoreLib.dasm (-0.34% of base)

1 total files with size differences (1 improved, 0 regressed), 0 unchanged.

Top method improvements by size (bytes):
       -2719 (-59.08% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,double,int):int
       -2672 (-56.13% of base) : System.Private.CoreLib.dasm - SpanHelpers:IndexOfAny(byref,double,double,double,int):int
       -1796 (-57.18% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,int):int
       -1793 (-57.28% of base) : System.Private.CoreLib.dasm - SpanHelpers:IndexOfAny(byref,double,double,int):int
        -869 (-52.51% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOf(byref,double,int):int

Top method improvements by size (percentage):
         -31 (-62.00% of base) : System.Private.CoreLib.dasm - Double:IsNaN(double):bool
         -75 (-60.98% of base) : System.Private.CoreLib.dasm - Double:Equals(double):bool:this
         -67 (-59.29% of base) : System.Private.CoreLib.dasm - GenericEqualityComparer`1:Equals(double,double):bool:this
       -2719 (-59.08% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,double,int):int
         -78 (-57.35% of base) : System.Private.CoreLib.dasm - Double:CompareTo(double):int:this
```

The JIT change on top of that gives:
```
Top file improvements by size (bytes):
       -1191 : System.Private.CoreLib.dasm (-0.03% of base)
1 total files with size differences (1 improved, 0 regressed), 0 unchanged.
Top method improvements by size (bytes):
        -246 (-11.78% of base) : System.Private.CoreLib.dasm - SpanHelpers:IndexOfAny(byref,double,double,double,int):int
        -202 (-10.73% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,double,int):int
        -134 (-9.96% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,int):int
        -130 (-9.72% of base) : System.Private.CoreLib.dasm - SpanHelpers:IndexOfAny(byref,double,double,int):int
         -78 (-1.70% of base) : System.Private.CoreLib.dasm - SpanHelpers:SequenceEqual(byref,byref,int):bool (7 methods)
Top method improvements by size (percentage):
          -5 (-26.32% of base) : System.Private.CoreLib.dasm - Double:IsNaN(double):bool
          -5 (-26.32% of base) : System.Private.CoreLib.dasm - Single:IsNaN(float):bool
        -246 (-11.78% of base) : System.Private.CoreLib.dasm - SpanHelpers:IndexOfAny(byref,double,double,double,int):int
          -5 (-10.87% of base) : System.Private.CoreLib.dasm - GenericEqualityComparer`1:Equals(double,double):bool:this
        -202 (-10.73% of base) : System.Private.CoreLib.dasm - SpanHelpers:LastIndexOfAny(byref,double,double,double,int):int
43 total methods with size differences (43 improved, 0 regressed), 20461 unchanged.
```

So the S.P.Corelib itself is great, but the additional JIT optimization still brings in some more nice changes.