@tannergooding 2019-10-17
This would likely also improve WPF/WinForms scenarios and would allow removal of their custom workarounds (such as: https://source.dot.net/#WindowsBase/DoubleUtil.cs,303).