@mikedn 2019-10-17
> by updating IsNaN(x) to just do x != x

Benchmarks? AFAIR I tried this in the past and the numbers weren't quite what I hoped.