@CarolEidt 2019-10-17
> I would expect that if both operands use the same variable they'll get the same register. @CarolEidt

Yes; it would be very surprising if they did not. The only case that should happen is when we have a call where one argument must be moved or copied to a different register.