@tannergooding 2019-10-17
> Benchmarks? AFAIR I tried this in the past and the numbers weren't quite what I hoped.

From a fairly simple benchmark (https://gist.github.com/tannergooding/102e20a899bdb6de6b4bcff38c61ee29), I'm seeing about a 25% perf increase:
```csharp
~\source\repos\ConsoleApp20\ConsoleApp20\bin\Release\netcoreapp3.0> C:\repos\coreclr_base\bin\tests\Windows_NT.x64.release\Tests\Core_Root\CoreRun.exe .\ConsoleApp20.dll                                                                       Elapsed: 2021.4188
Min: 0.9799
Max: 1.597
Avg: 1.0107093999999999

Elapsed: 2006.7065
Min: 0.998
Max: 1.5157
Avg: 1.00335325

~\source\repos\ConsoleApp20\ConsoleApp20\bin\Release\netcoreapp3.0> C:\repos\coreclr\bin\tests\Windows_NT.x64.release\Tests\Core_Root\CoreRun.exe .\ConsoleApp20.dll                                                                            Elapsed: 1601.1136
Min: 0.7887
Max: 1.3093
Avg: 0.8005568

Elapsed: 1572.2612
Min: 0.7819
Max: 1.1145
Avg: 0.7861306
```

I'm waiting to hear back from @adamsitnik on if there is an easy way to run Benchmark.NET against a private coreclr build...