@CarolEidt 2019-10-17
You can run Benchmark.NET against a private coreclr build, though you'll want to select a subset (there are a *lot* of benchmarks!) Here's a sample of how I run the micro benchmarks. From the performance\src\benchmarks\micro directory:
```
dotnet run -c Release -f netcoreapp5.0 --statisticalTest 3% --corerun E:\Test\EHWriteThruLSRA\Master\Core_Root\CoreRun.exe E:\Test\EHWriteThruLSRA\Mine\Core_Root\CoreRun.exe --filter *Foo*
```
Where `Foo` is a substring that you want to match.