@tannergooding 2019-10-17
Added a benchmark.net benchmark here: https://github.com/dotnet/performance/pull/952

Local results show:
```
| Method |                                                                  Toolchain | value |     Mean |   Error |  StdDev |   Median |      Min |      Max | Ratio | MannWhitney(3%) | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |--------------------------------------------------------------------------- |------ |---------:|--------:|--------:|---------:|---------:|---------:|------:|---------------- |------:|------:|------:|----------:|
|  IsNaN |      \coreclr\bin\tests\Windows_NT.x64.Release\Tests\Core_Root\CoreRun.exe |   NaN | 813.5 us | 0.46 us | 0.38 us | 813.3 us | 813.1 us | 814.5 us |  0.92 |          Faster |     - |     - |     - |         - |
|  IsNaN | \coreclr_base\bin\tests\Windows_NT.x64.Release\Tests\Core_Root\CoreRun.exe |   NaN | 881.4 us | 2.62 us | 2.45 us | 880.8 us | 877.7 us | 886.7 us |  1.00 |            Base |     - |     - |     - |         - |
|        |                                                                            |       |          |         |         |          |          |          |       |                 |       |       |       |           |
|  IsNaN |      \coreclr\bin\tests\Windows_NT.x64.Release\Tests\Core_Root\CoreRun.exe |     0 | 813.7 us | 0.30 us | 0.25 us | 813.7 us | 813.4 us | 814.3 us |  0.92 |          Faster |     - |     - |     - |         - |
|  IsNaN | \coreclr_base\bin\tests\Windows_NT.x64.Release\Tests\Core_Root\CoreRun.exe |     0 | 881.5 us | 2.55 us | 2.39 us | 882.0 us | 878.2 us | 886.0 us |  1.00 |            Base |     - |     - |     - |         - |
```